update Long to int

+++ src/main/java/com/ngo/fundraiser/entity/CampaignBeneficiaries.java
@@ -8,8 +8,7 @@ public class CampaignBeneficiaries {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long cBeneID;
-
+    private int cBeneID;
     // added for bi directional to campaigns table
     @ManyToOne
     @JoinColumn(name = "campaignID")
@@ -21,7 +20,7 @@ public class CampaignBeneficiaries {
 
     private int DonationValue;
 
-    public Long getCBeneID() {
+    public int getCBeneID() {
         return cBeneID;
     }
